// src/lib/types/enterprise.ts
export interface ScheduledFormType {
  email: string;
  firstName: string;
  lastName: string;
  mobile: string;
  countryCode: string;
  title: string;
  services: string[];
  country: string;
  locations: string[];
  description: string;
  startDate: string;
  endDate: string;
  currency: string;
  amount: string;
  rateUnit: string;
  level: string;
  expertId: string;
}